requirements.txt
pandas
numpy
scipy
pyyaml
openai
python-dotenv
openpyxl
shiny
ipykernel
jinja2
langchain
langchain-openai
chromadb
pypdf
langchain-text-splitters